+++ src/repetition.cpp
@@ -48,7 +48,6 @@ void Repetition::print() const {
         case RepetitionType::None:
             return;
     }
-    properties_print(properties);
 }
 
 void Repetition::clear() {
@@ -57,13 +56,11 @@ void Repetition::clear() {
     } else if (type == RepetitionType::ExplicitX || type == RepetitionType::ExplicitY) {
         coords.clear();
     }
-    properties_clear(properties);
     memset(this, 0, sizeof(Repetition));
 }
 
 void Repetition::copy_from(const Repetition repetition) {
     type = repetition.type;
-    properties = properties_copy(repetition.properties);
     switch (type) {
         case RepetitionType::Rectangular:
             columns = repetition.columns;