+++ src/repetition.h
@@ -41,7 +41,6 @@ struct Repetition {
         Array<Vec2> offsets;   // Explicit (10, 11)
         Array<double> coords;  // ExplicitX, ExplicitY (Oasis 4, 5, 6, 7)
     };
-    Property* properties;
 
     void print() const;
     void clear();